backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

